New rule based tagger outputs token indexes of MWE
