+++ pymusas/taggers/new_rule_based.py
@@ -11,7 +11,7 @@ class RuleBasedTagger():
     to tag will apply one or more :class:`pymusas.taggers.rules.rule.Rule`s
     to create a list of possible candidate tags for each token in the sequence.
     Each candidate, represented as a
-    :class:`pymusas.rankers.lexicon_entry.RankingMetaData` object, for each
+    :class:`pymusas.rankers.ranking_meta_data.RankingMetaData` object, for each
     token is then Ranked using a
     :class:`pymusas.rankers.lexicon_entry.LexiconEntryRanker` ranker. The best
     candidate and it's associated tag(s) for each token are then returned along
@@ -72,8 +72,8 @@ class RuleBasedTagger():
         2. A `List` of `Tuples` whereby each `Tuple` indicates the start and end
         token index of the associated Multi Word Expression (MWE). If the `List` contains
         more than one `Tuple` then the MWE is discontinuous. For single word
-        expressions the `List` will only contain 1 `Tuple` which will only contain
-        an index of (start_index, start_index + 1).
+        expressions the `List` will only contain 1 `Tuple` which will be
+        (token_start_index, token_start_index + 1).
     
         All the generated tags and MWEs are based on the rules and ranker given
         to this model.
