+++ pymusas/taggers/rules/rule.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import List
 
-from pymusas.rankers.lexicon_entry import RankingMetaData
+from pymusas.rankers.ranking_meta_data import RankingMetaData
 
 
 class Rule(ABC):
@@ -11,7 +11,7 @@ class Rule(ABC):
 
     A Rule when called, `__call__`, creates a `List` of rules matches for each
     token, whereby each rule matched is defined by the
-    :class:`pymusas.rankers.lexicon_entry.RankingMetaData` object. These
+    :class:`pymusas.rankers.ranking_meta_data.RankingMetaData` object. These
     rules matches per token can then be, optionally, combined with other rule
     matches per token from other :class:`Rule` classes to then be ranked by a
     :class:`pymusas.rankers.lexicon_entry.LexiconEntryRanker`.
@@ -22,7 +22,7 @@ class Rule(ABC):
                  pos_tags: List[str]) -> List[List[RankingMetaData]]:
         '''
         For each token it returns a `List` of rules matches defined by the
-        :class:`pymusas.rankers.lexicon_entry.RankingMetaData` object.
+        :class:`pymusas.rankers.ranking_meta_data.RankingMetaData` object.
 
         Each `List` of `tokens`, `lemmas`, and `pos_tags` are assumed to be of
         equal length.
